fix doubled directory and merge changes into base
