--- packages/react-drag-drop/src/components/DragDrop/DroppableContext.ts
@@ -1,6 +1,0 @@
-import * as React from 'react';
-
-export const DroppableContext = React.createContext({
-  zone: 'defaultDroppableZone',
-  droppableId: 'defaultDroppableId'
-});
