+++ packages/react-drag-drop/src/components/DragDrop/__tests__/DragDrop.test.tsx
@@ -2,6 +2,11 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { DragDropSort } from '../';
 
+jest.mock('react-dom', () => ({
+  ...jest.requireActual('react-dom'),
+  createPortal: jest.fn((node) => node)
+}));
+
 test('renders some divs', () => {
   const { asFragment } = render(
     <div className="pf-c-droppable pf-m-dragging">
