+++ packages/react-drag-drop/src/components/DragDrop/index.ts
@@ -1,1 +1,3 @@
 export * from './DragDropSort';
+export * from './Droppable';
+export * from './DragDropContainer';
