--- packages/react-drag-drop/src/next/components/DragDrop/DragButton.tsx
@@ -1,27 +1,0 @@
-import * as React from 'react';
-import { css } from '@patternfly/react-styles';
-import dragButtonStyles from '@patternfly/react-styles/css/components/DataList/data-list';
-import buttonStyles from '@patternfly/react-styles/css/components/Button/button';
-import GripVerticalIcon from '@patternfly/react-icons/dist/esm/icons/grip-vertical-icon';
-
-export interface DragButtonProps extends React.HTMLProps<HTMLButtonElement> {
-  /** Additional classes added to the drag button */
-  className?: string;
-  /** Sets button type */
-  type?: 'button' | 'submit' | 'reset';
-  /** Flag indicating if drag is disabled for the item */
-  isDisabled?: boolean;
-}
-
-export const DragButton: React.FunctionComponent<DragButtonProps> = ({ className, ...props }: DragButtonProps) => (
-  <button
-    className={css(className, buttonStyles.button, buttonStyles.modifiers.plain)}
-    aria-label="Drag button"
-    {...props}
-  >
-    <span className={css(dragButtonStyles.dataListItemDraggableIcon)}>
-      <GripVerticalIcon />
-    </span>
-  </button>
-);
-DragButton.displayName = 'DragButton';
