--- packages/react-drag-drop/src/next/components/DragDrop/index.ts
@@ -1,3 +1,0 @@
-export * from './DragDropSort';
-export * from './Droppable';
-export * from './DragDropContainer';
