--- packages/react-drag-drop/src/next/components/index.ts
@@ -1,1 +1,0 @@
-export * from './DragDrop';
