--- packages/react-drag-drop/src/next/index.ts
@@ -1,1 +1,0 @@
-export * from './components';
